Move onKeyDown logic into another file
